Clean commit history
Update ODataResourceSetSerializer.cs

Change the interface for skiptoken

Fix build issue

update filename

Don't have a true in the query itself

Update Core version to use abstract class for DI

Remove ISkipTokenConsumer

Fix Classic version
