+++ src/Microsoft.AspNet.OData.Shared/Formatter/Serialization/ODataSerializerContext.cs
@@ -5,10 +5,10 @@ using System;
 using System.Collections.Generic;
 using Microsoft.AspNet.OData.Common;
 using Microsoft.AspNet.OData.Interfaces;
+using Microsoft.AspNet.OData.Query;
 using Microsoft.OData.Edm;
 using ODataPath = Microsoft.AspNet.OData.Routing.ODataPath;
 using SelectExpandClause = Microsoft.OData.UriParser.SelectExpandClause;
-using OrderByClause = Microsoft.OData.UriParser.OrderByClause;
 
 namespace Microsoft.AspNet.OData.Formatter.Serialization
 {
@@ -39,17 +39,18 @@ namespace Microsoft.AspNet.OData.Formatter.Serialization
         public ODataSerializerContext(ResourceContext resource, SelectExpandClause selectExpandClause, IEdmProperty edmProperty)
             : this(resource, selectExpandClause, edmProperty, null)
         {
-
         }
 
         /// <summary>
-        /// 
+        /// Initializes a new instance of the <see cref="ODataSerializerContext"/> class.
         /// </summary>
-        /// <param name="resource"></param>
-        /// <param name="selectExpandClause"></param>
-        /// <param name="orderByClause"></param>
-        /// <param name="edmProperty"></param>
-        public ODataSerializerContext(ResourceContext resource, SelectExpandClause selectExpandClause, IEdmProperty edmProperty, OrderByClause orderByClause)
+        /// <param name="resource">The resource whose navigation property is being expanded.</param>
+        /// <param name="selectExpandClause">The <see cref="SelectExpandClause"/> for the navigation property being expanded.</param>
+        /// <param name="edmProperty">The complex property being nested or the navigation property being expanded.
+        /// If the resource property is the dynamic complex, the resource property is null.
+        /// </param>
+        /// <param name="queryOptions">The <see cref="ODataQueryOptions"/> for the navigation property being expanded.</param>
+        public ODataSerializerContext(ResourceContext resource, SelectExpandClause selectExpandClause, IEdmProperty edmProperty, ODataQueryOptions queryOptions)
         {
             if (resource == null)
             {
@@ -70,7 +71,7 @@ namespace Microsoft.AspNet.OData.Formatter.Serialization
 
             ExpandedResource = resource; // parent resource
             SelectExpandClause = selectExpandClause;
-            OrderByClause = orderByClause;
+            QueryOptions = queryOptions;
             EdmProperty = edmProperty; // should be nested property
 
             if (context.NavigationSource != null)
@@ -126,9 +127,9 @@ namespace Microsoft.AspNet.OData.Formatter.Serialization
         public SelectExpandClause SelectExpandClause { get; set; }
 
         /// <summary>
-        /// Gets or sets the <see cref="SelectExpandClause"/>.
+        /// Gets or sets the <see cref="ODataQueryOptions"/>.
         /// </summary>
-        public OrderByClause OrderByClause { get; set; }
+        public ODataQueryOptions QueryOptions { get; set; }
 
         /// <summary>
         /// Gets or sets the resource that is being expanded.
