+++ src/Microsoft.AspNet.OData/GlobalSuppressions.cs
@@ -37,4 +37,4 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Microsoft.Globalization", "CA1307:SpecifyStringComparison", MessageId = "System.String.StartsWith(System.String)", Scope = "member", Target = "Microsoft.AspNet.OData.Query.DefaultSkipTokenImplementation.#ProcessSkipTokenValue(System.String)")]
 [assembly: SuppressMessage("Microsoft.Design", "CA1020:AvoidNamespacesWithFewTypes", Scope = "namespace", Target = "Microsoft.AspNet.OData.Interfaces")]
 [assembly: SuppressMessage("Microsoft.Maintainability", "CA1506:AvoidExcessiveClassCoupling", Scope = "member", Target = "Microsoft.AspNet.OData.Formatter.Serialization.ODataResourceSetSerializer.#CreateResourceSet(System.Collections.IEnumerable,Microsoft.OData.Edm.IEdmCollectionTypeReference,Microsoft.AspNet.OData.Formatter.Serialization.ODataSerializerContext)")]
-
+[assembly: SuppressMessage("Microsoft.Globalization", "CA1307:SpecifyStringComparison", MessageId = "System.String.StartsWith(System.String)", Scope = "member", Target = "Microsoft.AspNet.OData.Query.DefaultSkipTokenHandler.#ProcessSkipTokenValue(System.String)")]
